// app_dashboard/src/services/blue-delta-sdk/model/Jean.ts
/**
 * BlueDelta-API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Jean {
    jeanId: string;

    measurement?: models.Measurement;

    thread?: models.Thread;

    fabric?: models.Fabric;

    button?: models.Button;

}